% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsc_knit.R
\name{bsc_knit}
\alias{bsc_knit}
\title{Define custom knit behavior}
\usage{
bsc_knit(input, ...)
}
\arguments{
\item{input}{The document to knit}

\item{...}{Additional arguments}
}
\description{
Define custom knit behavior
}
