% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_render.R
\name{coalesce_lists}
\alias{coalesce_lists}
\title{Coalesce two lists by element name}
\usage{
coalesce_lists(x, y)
}
\arguments{
\item{x}{A list. Values will take precedence if names overlap.}

\item{y}{A list. Values will be overwritten if names overlap.}
}
\value{
A list
}
\description{
Coalesce two lists by element name
}
