% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_render.R
\name{copy_dir}
\alias{copy_dir}
\title{Copy the contents of a file from one location to another}
\usage{
copy_dir(from, to)
}
\arguments{
\item{from}{A directory path}

\item{to}{A directory path}
}
\description{
Copy the contents of a file from one location to another
}
