% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_for_web.R
\name{copy_for_web}
\alias{copy_for_web}
\title{Copy output from one location to another (intended to pipe locally from bscContent to bsc-site)}
\usage{
copy_for_web(bundle_name, bundle_path, dest_path, clean = TRUE)
}
\arguments{
\item{bundle_name}{The name of the page bundle (e.g., "e_consent")}

\item{bundle_path}{The path to the page bundle WITHOUT the bundle name (e.g., "~/project/output")}

\item{dest_path}{The path to the destination of the page bundle (e.g., "~/site/section")}

\item{clean}{Recursively delete everything in "~/site/section/bundle_name" before copying?}
}
\description{
Copy output from one location to another (intended to pipe locally from bscContent to bsc-site)
}
