% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_shared.R
\name{copy_from_shared}
\alias{copy_from_shared}
\title{Copy files from a shared location; optionally overwrite YAML}
\usage{
copy_from_shared(old_dir, new_dir, new_yaml = NULL, log = "log.md")
}
\arguments{
\item{old_dir}{The old location (full path to directory)}

\item{new_dir}{The new location (full path to directory)}

\item{new_yaml}{New YAML for the Rmd file. Will overwrite old YAML.}

\item{log}{(If not NULL) Relative path to log file}
}
\description{
Copy files from a shared location; optionally overwrite YAML
}
