% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_for_web.R
\name{create_attachment_links}
\alias{create_attachment_links}
\title{Create markdown links to a list of files in a directory}
\usage{
create_attachment_links(
  source_dir = "files",
  target_dir = source_dir,
  pattern = NULL,
  exclude = NULL
)
}
\arguments{
\item{source_dir}{Where are the files located?}

\item{target_dir}{Where should references point? (probably the same as \code{source_dir})}

\item{pattern}{Regex search string. Defaults to NULL (returns everything).}

\item{exclude}{Specific files to exclude from the list. Only basenames are compared.}
}
\value{
Vector of strings, suitable to be inserted into text as a bulleted list
}
\description{
Create markdown links to a list of files in a directory
}
