% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_render.R
\name{create_temp_bundle}
\alias{create_temp_bundle}
\title{Copy files from a shared location; optionally overwrite YAML}
\usage{
create_temp_bundle(
  input,
  default_yaml = pkg_resource("rmd_files/web_defaults.yml"),
  pre_content = c(pkg_resource("rmd_files/web_opts_chunk.Rmd"),
    pkg_resource("rmd_files/web_page_toc.Rmd")),
  post_content = c(pkg_resource("rmd_files/web_reference_block.Rmd")),
  temp_dir = tempdir()
)
}
\arguments{
\item{input}{A path to a .Rmd file.}

\item{default_yaml}{A path to a .yml file containing default key-value pairs. Values will be overwritten by matching values in input header.}

\item{pre_content}{A path to a file containing text to be inserted after the YAML header and before existing input content.}

\item{post_content}{A path to a file containing text to be inserted after existing input content.}

\item{temp_dir}{A path where the modified page bundle (directory) will be created}
}
\description{
Copy files from a shared location; optionally overwrite YAML
}
