% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_shared.R
\name{default_opts_chunk}
\alias{default_opts_chunk}
\title{Return the default chunk options for a BSC doc}
\usage{
default_opts_chunk()
}
\value{
A vector of strings
}
\description{
Return the default chunk options for a BSC doc
}
