% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gd_id.R
\name{find_gd_id}
\alias{find_gd_id}
\title{Look Up the Id of a Google Drive Subdirectory}
\usage{
find_gd_id(dirname, type = "folder", gd_id = Sys.getenv("gd_id"))
}
\arguments{
\item{dirname}{The name of the subdirectory. Can be a partial match}

\item{type}{Type of resource to find. Defaults to "folder".}

\item{gd_id}{The id of the parent folder on Google Drive. Defaults to the environmental variable \code{gd_id}.}
}
\value{
The Google Drive id (a string).
}
\description{
Look Up the Id of a Google Drive Subdirectory
}
\examples{
\dontrun{find_gd_id("templates")}
}
