% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_theme_doc.R
\name{find_theme_doc}
\alias{find_theme_doc}
\title{Fetch the reference doc associated with a theme}
\usage{
find_theme_doc(theme, extension)
}
\arguments{
\item{theme}{The name of the theme. Must match the name of the reference file.}

\item{extension}{The extension of the reference doc, without a period}
}
\value{
The path to the reference document. If no document was found, returns \code{NULL}.
}
\description{
Fetch the reference doc associated with a theme
}
\examples{
find_theme_doc("red", extension = "pptx")
}
