% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_shared.R
\name{geekdoc_toc}
\alias{geekdoc_toc}
\title{Return a chunk that contains a geekdoc TOC}
\usage{
geekdoc_toc()
}
\value{
A vector of strings
}
\description{
Return a chunk that contains a geekdoc TOC
}
