% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_draft.R
\name{html_draft}
\alias{html_draft}
\title{HTML document output format, with optional themes}
\usage{
html_draft(toc = TRUE, theme = "plain")
}
\arguments{
\item{toc}{Include a table of contents (2 levels deep)?}

\item{theme}{Name of theme used in styling the output}
}
\description{
HTML document output format, with optional themes
}
