% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_tipsheet.R
\name{knit_tipsheet}
\alias{knit_tipsheet}
\title{Define Knitting Behavior for a Tipsheet}
\usage{
knit_tipsheet(
  input,
  output_dir = NULL,
  output_file = NULL,
  date_in_name = TRUE,
  ...
)
}
\arguments{
\item{input}{The (current) input file}

\item{output_dir}{The location of the output file. Defaults to the same location as the Rmd file.}

\item{output_file}{The name of the output file, without an extension. Defaults to the input file name.}

\item{date_in_name}{Logical. Should the output file name include today's date?}

\item{...}{Additional arguments}
}
\description{
Define Knitting Behavior for a Tipsheet
}
