% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_bib.R
\name{latex_to_katex}
\alias{latex_to_katex}
\title{LaTeX to KaTeX}
\usage{
latex_to_katex(text)
}
\arguments{
\item{text}{A string or vector of strings}
}
\value{
A string or vector of strings with html-safe math symbols
}
\description{
LaTeX to KaTeX
}
