% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_render.R
\name{modify_rmd_contents}
\alias{modify_rmd_contents}
\title{Insert YAML values, headers, and footers in an existing Rmd file.}
\usage{
modify_rmd_contents(
  input,
  default_yaml = NULL,
  pre_content = NULL,
  post_content = NULL
)
}
\arguments{
\item{input}{A path to a .Rmd file.}

\item{default_yaml}{A path to a .yml file containing default key-value pairs. Values will be overwritten by matching values in input header.}

\item{pre_content}{A path to a file containing text to be inserted after the YAML header and before existing input content.}

\item{post_content}{A path to a file containing text to be inserted after existing input content.}
}
\value{
a vector of strings
}
\description{
Insert YAML values, headers, and footers in an existing Rmd file.
}
