% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_document.R
\name{pdf_document}
\alias{pdf_document}
\title{PDF Document}
\usage{
pdf_document(toc = FALSE, toc_depth = NULL, number_sections = FALSE, ...)
}
\arguments{
\item{toc}{Table of contents?}

\item{toc_depth}{Numeric. Level of headings included in table of contents}

\item{number_sections}{Logical. If \code{TRUE}, sections will be numbered according to heading hierarchy}

\item{...}{Additional arguments passed to bookdown::pdf_document2()}
}
\description{
PDF Document
}
