% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_resource.R
\name{pkg_resource}
\alias{pkg_resource}
\title{Fetch a Resource from the Current Package}
\usage{
pkg_resource(..., package = "bscContentHelpers")
}
\arguments{
\item{...}{The search string for the desired resource}

\item{package}{The name of the package}
}
\value{
The path to the resource
}
\description{
Fetch a Resource from the Current Package
}
\examples{
pkg_resource("rmd_files/custom.css")
}
