% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_presentation.R
\name{pptx_presentation}
\alias{pptx_presentation}
\title{Powerpoint (pptx) document output format, with optional themes}
\usage{
pptx_presentation(theme = "red")
}
\arguments{
\item{theme}{Name of theme to use when styling output presentation}
}
\description{
Powerpoint (pptx) document output format, with optional themes
}
