% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_shared.R
\name{prepend_c}
\alias{prepend_c}
\title{Utility function to prepend items in a vector through the pipe}
\usage{
prepend_c(x, ...)
}
\arguments{
\item{x}{A value or vector}

\item{...}{Values to add to the beginning of the vector}
}
\value{
A vector
}
\description{
Utility function to prepend items in a vector through the pipe
}
