% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_bib.R
\name{show_bib}
\alias{show_bib}
\title{Show the contents of a BibTeX file}
\usage{
show_bib(name, dir = "files", html = knitr::is_html_output())
}
\arguments{
\item{name}{The basename of the .bib file}

\item{dir}{The location of the .bib file. Absolute or relative to
the working directory.}

\item{html}{Should the output be made html-safe?}
}
\value{
A formatted list of references. To show in an
RMarkdown document, set the chunk option \code{results='asis'}.
}
\description{
Return a formatted list of references in a .bib file, probably in
the files/ subdirectory, without citing them.#'
}
