% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_shared.R
\name{strip_yaml}
\alias{strip_yaml}
\title{Return only the content of a markdown file, sans YAML}
\usage{
strip_yaml(x)
}
\arguments{
\item{x}{A vector of strings (probably from readLines)}
}
\value{
The vector without any text between '---' delimiters
}
\description{
Return only the content of a markdown file, sans YAML
}
