% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_render.R
\name{unyaml_file}
\alias{unyaml_file}
\title{Return only the content of a markdown file, sans YAML}
\usage{
unyaml_file(file)
}
\arguments{
\item{file}{A path to a file with a YAML header}
}
\value{
A vector of strings without yaml header
}
\description{
Return only the content of a markdown file, sans YAML
}
