% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_shared.R
\name{update_yaml}
\alias{update_yaml}
\title{Update Existing YAML}
\usage{
update_yaml(file, new_yaml = NULL)
}
\arguments{
\item{file}{A file with a YAML header}

\item{new_yaml}{Fields to add or overwrite}
}
\value{
A string with new YAML values
}
\description{
Update Existing YAML
}
