% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_document.R
\name{word_document}
\alias{word_document}
\title{Word (docx) document output format, with optional themes}
\usage{
word_document(toc = TRUE, theme = "teal")
}
\arguments{
\item{toc}{Include a table of contents (2 levels deep)?}

\item{theme}{Theme to use in styling the output document}
}
\description{
Word (docx) document output format, with optional themes
}
