% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_draft.R
\name{html_draft}
\alias{html_draft}
\title{Knit a Document to an HTML Draft}
\usage{
html_draft(toc = TRUE, css = NULL)
}
\arguments{
\item{toc}{Include a table of contents (2 levels deep)?}

\item{css}{Path to css stylesheet}
}
\description{
Knit a Document to an HTML Draft
}
