% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_render.R
\name{knit_for_web}
\alias{knit_for_web}
\title{Prepare a Rmd file and supporting files for inclusion in a blogdown project}
\usage{
knit_for_web(input, target_dir, ..., clean = FALSE)
}
\arguments{
\item{input}{A path to the input file}

\item{target_dir}{A path to the directory where the page bundle should be located}

\item{...}{Additional arguments passed to create_temp_bundle()}

\item{clean}{Devele everything in target_dir before populating?}
}
\description{
Prepare a Rmd file and supporting files for inclusion in a blogdown project
}
